patches
- changed qureg's rank to be 0 on every node when qureg is not distributed, even when the env is distributed, so that the many backend routines which consult rank (and are intended agnostic to distribution settings) are correct
- patched probability calculations for density matrices, which erroneously re-leveraged the statevector routine 
- patched decoherence validation which, by trying to permit prob=0 decoherence on statevectors, actually ran the routines on statevectors
- added qureg superposition validation missing from previous commit
- added definition for syncQuESTEnv, for developer convenience

Note that on Victor's Quadro P6000, `cuquantum_statevec_anyCtrlAnyTargDenseMatrix_subA()` is not modifying the state. I cannot determine why; it may be due to the P6000 having an insufficient compute-capability for cuStateVec (it is 6.2, below the minimum 7.5), so we defer the issue

+++ quest/src/gpu/gpu_cuquantum.cuh
@@ -294,8 +294,9 @@ void cuquantum_densmatr_twoQubitDephasing_subA(Qureg qureg, int qubitA, int qubi
 
 qreal cuquantum_statevec_calcTotalProb_sub(Qureg qureg) {
 
-    qreal prob0;
-    qreal prob1;
+    // cuQuantum probabilities are always double (not qreal)
+    double prob0;
+    double prob1;
 
     // find probablity of leftmost qubit, so that reduction is 
     // contiguous, which we expect has the optimum performance
@@ -314,15 +315,16 @@ qreal cuquantum_statevec_calcTotalProb_sub(Qureg qureg) {
 
 qreal cuquantum_statevec_calcProbOfMultiQubitOutcome_sub(Qureg qureg, vector<int> qubits, vector<int> outcomes) {
 
-    qreal prob0;
-    qreal* prob1 = nullptr; // don't compute 1-prob0, because we assume normalisation
+    // cuQuantum probabilities are always double (not qreal)
+    double prob0;
+    double* prob1 = nullptr; // don't compute 1-prob0, because we assume normalisation
 
     CUDA_CHECK( custatevecAbs2SumOnZBasis(
         config.cuQuantumHandle,
         toCuQcomps(qureg.gpuAmps), CUQUANTUM_QCOMP, qureg.logNumAmpsPerNode,
         &prob0, prob1, qubits.data(), qubits.size() ) );
 
-    return prob0;
+    return (qreal) prob0;
 }
